refactor: Co-locate used ABIs inside of SDK - Reduces package size - Allows for ABIs to change over time
